Create mapping.json with my answer
Had browsed the content related to the "autofill detail token" in the page (https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#attr-fe-autocomplete), and actually not quite understand with the question, I supposed maybe the answer would be the metadata hash which return by Task #2. So used "console.info(JSON.stringify(hash))" and copy the hash from the chrome console. And the question spec requests to "formatted and indented", so added \n manually.